requirements.txt
alembic==0.8.8
appdirs==1.4.0
Babel==2.3.4
Beaker==1.8.0
bleach==1.4.3
bleach-whitelist==0.0.7
Chameleon==2.25
colander==1.2
deform==2.0a2
docopt==0.6.2
fanstatic==1.0a7
filedepot==0.4.0
FormEncode==2.0.0a1
funcsigs==1.0.2
html2text==2016.9.19
html5lib==0.9999999
iso8601==0.1.11
js.angular==1.1.4
js.bootstrap==3.3.4
js.deform==2.0a2.post3
js.fineuploader==3.3.0
js.html5shiv==3.7.3
js.jquery==1.9.1
js.jquery-form==3.9
js.jquery-maskedinput==1.3.1
js.jquery-maskmoney==1.4.1
js.jquery-sortable==0.9.12
js.jquery-tablednd==0.4
js.jquery-timepicker-addon==1.5.3
js.jqueryui==1.10.3
js.jqueryui-tagit==2.0.24.post2
js.modernizr==2.5.3.1
js.select2==3.4.1
js.tinymce==4.2.7
kotti-image==0.1.4
kotti-tinymce==0.5.4
LEPL==5.1.3
lingua==4.10
Mako==1.0.4
MarkupSafe==0.23
PasteDeploy==1.5.2
peppercorn==0.5
polib==1.0.7
py-bcrypt==0.4
pyramid==1.7.3
pyramid-beaker==0.8
pyramid-chameleon==0.3
pyramid-deform==0.2
pyramid-mailer==0.14.1
pyramid-tm==1.0
pyramid-zcml==1.0.0
python-editor==1.0.1
pytz==2016.7
repoze.lru==0.6
repoze.sendmail==4.2
repoze.workflow==1.0b1
repoze.zcml==1.0b1
rfc6266==0.0.4
shutilwhich==1.1.0
six==1.10.0
SQLAlchemy==1.1.1
SQLAlchemy-Utils==0.32.9
transaction==1.6.1
translationstring==1.3
Unidecode==0.4.19
usersettings==1.0.7
venusian==1.0
waitress==1.0.0
WebOb==1.6.1
zope.component==4.3.0
zope.configuration==4.0.3
zope.deprecation==4.1.2
zope.event==4.2.0
zope.i18nmessageid==4.0.3
zope.interface==4.3.2
zope.schema==4.4.2
zope.sqlalchemy==0.7.7
